Moved the hystris stream event DTO to resilence4j common module
