+++ resilience4j-framework-common/src/main/java/io/github/resilience4j/common/circuitbreaker/monitoring/endpoint/CircuitBreakerHystrixStreamEventsDTO.java
@@ -1,29 +1,19 @@
-/*
- * Copyright 2020 Vijay Ram
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+package io.github.resilience4j.common.circuitbreaker.monitoring.endpoint;
 
-package io.github.resilience4j.ratpack.circuitbreaker.monitoring.endpoint.events;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
 import io.github.resilience4j.circuitbreaker.CircuitBreaker;
 import io.github.resilience4j.circuitbreaker.CircuitBreakerConfig;
 import io.github.resilience4j.circuitbreaker.event.CircuitBreakerEvent;
 
+/**
+ * @author vijayram
+ */
 
-@JsonInclude(JsonInclude.Include.NON_NULL)
-public class CircuitBreakerStreamEventsDTO {
+/**
+ * This class is used to emit available data as SSE so that resilence4j SSE can be bridged
+ * to suppose legacy hystrix eco tools such as hystrix dashboard , turbine and vizceral.
+ */
+
+public class CircuitBreakerHystrixStreamEventsDTO {
 
     private CircuitBreakerEvent circuitBreakerRecentEvent;
     private CircuitBreaker.Metrics metrics;
@@ -31,10 +21,10 @@ public class CircuitBreakerStreamEventsDTO {
     private float failureRateThreshold;
     private float slowCallRateThreshold;
 
-    public CircuitBreakerStreamEventsDTO(CircuitBreakerEvent circuitBreakerEvent,
-                                         CircuitBreaker.State state,
-                                         CircuitBreaker.Metrics metrics,
-                                         CircuitBreakerConfig circuitBreakerConfig) {
+    public CircuitBreakerHystrixStreamEventsDTO(CircuitBreakerEvent circuitBreakerEvent,
+                                                CircuitBreaker.State state,
+                                                CircuitBreaker.Metrics metrics,
+                                                CircuitBreakerConfig circuitBreakerConfig) {
         this.circuitBreakerRecentEvent = circuitBreakerEvent;
         this.metrics = metrics;
         this.currentState = state;
